Move google pay models to subpackage within model package
