+++ ryft-ui/src/main/java/com/ryftpay/android/ui/viewmodel/GooglePayResultViewModel.kt
@@ -3,7 +3,7 @@ package com.ryftpay.android.ui.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.ryftpay.android.ui.model.GooglePayResult
+import com.ryftpay.android.ui.model.googlepay.GooglePayResult
 
 internal class GooglePayResultViewModel : ViewModel() {
